+++ eyes_robotframework/src/EyesLibrary/config_parser.py
@@ -6,10 +6,11 @@ from enum import Enum
 from typing import Callable, Text, Type
 
 import trafaret as trf
-import yaml
 from six import raise_from
 
 from applitools.common import (
+    AndroidDeviceInfo,
+    AndroidDeviceName,
     ChromeEmulationInfo,
     DesktopBrowserInfo,
     DeviceName,
@@ -39,6 +40,7 @@ class SelectedRunner(Enum):
     web = "web"
     web_ufg = "web_ufg"
     mobile_native = "mobile_native"
+    native_mobile_grid = "native_mobile_grid"
 
 
 class RobotStitchMode(Enum):
@@ -158,6 +160,26 @@ class IosDeviceInfoTrafaret(trf.Trafaret):
         return [IosDeviceInfo(**dct) for dct in sanitized]
 
 
+class AndroidDeviceInfoTrafaret(trf.Trafaret):
+    scheme = trf.List(
+        trf.Dict(
+            {
+                "device_name": TextToEnumTrafaret(AndroidDeviceName),
+                trf.Key("screen_orientation", optional=True): UpperTextToEnumTrafaret(
+                    ScreenOrientation
+                ),
+                trf.Key("android_version", optional=True): UpperTextToEnumTrafaret(
+                    IosVersion
+                ),
+            }
+        )
+    )
+
+    def check_and_return(self, value, context=None):
+        sanitized = self.scheme.check(value, context)
+        return [AndroidDeviceInfo(**dct) for dct in sanitized]
+
+
 class ChromeEmulationInfoTrafaret(trf.Trafaret):
     scheme = trf.List(
         trf.Dict(
@@ -216,6 +238,21 @@ class BrowsersTrafaret(trf.Trafaret):
         return list(itertools.chain(desktop, ios, chrome_emulation))
 
 
+class DevicesTrafaret(trf.Trafaret):
+    scheme = trf.Dict(
+        {
+            trf.Key("ios", optional=True): IosDeviceInfoTrafaret,
+            trf.Key("android", optional=True): AndroidDeviceInfoTrafaret,
+        }
+    )
+
+    def check_and_return(self, value, context=None):
+        sanitized = self.scheme.check(value, context)
+        ios = sanitized.pop("ios", [])
+        android = sanitized.pop("android", [])
+        return list(itertools.chain(ios, android))
+
+
 class ConfigurationTrafaret(trf.Trafaret):  # typedef
     shared_scheme = trf.Dict(
         {
@@ -237,9 +274,10 @@ class ConfigurationTrafaret(trf.Trafaret):  # typedef
             trf.Key("properties", optional=True): trf.List(
                 trf.Dict(name=trf.String, value=trf.String)
             ),
+            trf.Key("wait_before_capture", optional=True): trf.Int,
         },
     )
-    selenium_scheme = shared_scheme + trf.Dict(
+    web_scheme = shared_scheme + trf.Dict(
         {
             trf.Key("force_full_page_screenshot", optional=True): trf.Bool,
             trf.Key("wait_before_screenshots", optional=True): trf.Int,
@@ -250,7 +288,7 @@ class ConfigurationTrafaret(trf.Trafaret):  # typedef
             trf.Key("hide_caret", optional=True): trf.Bool,
         },
     )
-    selenium_ufg_scheme = shared_scheme + trf.Dict(
+    web_ufg_scheme = shared_scheme + trf.Dict(
         {
             trf.Key("runner_options", optional=True): RunnerOptionsTrafaret,
             trf.Key("visual_grid_options", optional=True): VisualGridOptionsTrafaret,
@@ -261,16 +299,24 @@ class ConfigurationTrafaret(trf.Trafaret):  # typedef
             trf.Key("browsers", optional=True) >> "_browsers_info": BrowsersTrafaret,
         },
     )
-    appium_scheme = shared_scheme + trf.Dict(
+    mobile_native_scheme = shared_scheme + trf.Dict(
         {
             trf.Key("is_simulator", optional=True): trf.Bool,
         },
     )
+    native_mobile_grid_scheme = shared_scheme + trf.Dict(
+        {
+            trf.Key("devices", optional=True) >> "_browsers_info": DevicesTrafaret,
+        }
+    )
     scheme = shared_scheme + trf.Dict(
         {
-            trf.Key("web", optional=True): shared_scheme + selenium_scheme,
-            trf.Key("web_ufg", optional=True): shared_scheme + selenium_ufg_scheme,
-            trf.Key("mobile_native", optional=True): shared_scheme + appium_scheme,
+            trf.Key("web", optional=True): shared_scheme + web_scheme,
+            trf.Key("web_ufg", optional=True): shared_scheme + web_ufg_scheme,
+            trf.Key("mobile_native", optional=True): shared_scheme
+            + mobile_native_scheme,
+            trf.Key("native_mobile_grid", optional=True): shared_scheme
+            + native_mobile_grid_scheme,
         },
     )
 
@@ -288,6 +334,7 @@ class ConfigurationTrafaret(trf.Trafaret):  # typedef
         combined_raw_config.pop(SelectedRunner.web.value)
         combined_raw_config.pop(SelectedRunner.web_ufg.value)
         combined_raw_config.pop(SelectedRunner.mobile_native.value)
+        combined_raw_config.pop(SelectedRunner.native_mobile_grid.value)
 
         # add config for selected runner
         combined_raw_config.update(selected_sdk_conf)