+++ tests/functional/eyes_selenium/visual_grid/test_ufg_native.py
@@ -36,6 +36,10 @@ def test_ufg_native_ios_basic(sauce_driver_url):
         eyes.check_window()
         eyes.close()
 
+        all_results = runner.get_all_test_results()
+
+        assert all_results.passed == 1
+
 
 @pytest.mark.skip("Skip until test working build of test app is available")
 def test_ufg_android_basic(sauce_driver_url):
@@ -61,3 +65,7 @@ def test_ufg_android_basic(sauce_driver_url):
         eyes.open(driver, "USDK Test", "UFG native Android basic test")
         eyes.check_window()
         eyes.close()
+
+        all_results = runner.get_all_test_results()
+
+        assert all_results.passed == 1