+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Added
+- [eyes-robotframework] UFG for mobile native apps [Trello 2863](https://trello.com/c/63ARcWd5)
+
 ## [5.4.0] - 2022-04-13
 ### Added
 - [eyes-robotframework] Configuration file as environment variable [Trello 2768](https://trello.com/c/cTTD36eN)